requirement.txt
asgiref==3.4.1
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==8.0.1
cryptography==3.4.7
decorator==5.0.9
dj-database-url==0.5.0
Django==3.2.6
django-cors-headers==3.7.0
django-heroku==0.3.1
django-rest-knox==4.1.0
djangorestframework==3.12.4
dynaconf==3.1.4
gunicorn==20.1.0
hvac==0.11.0
idna==3.2
knox==0.1.14
loguru==0.5.3
psycopg2==2.9.1
psycopg2-binary==2.9.1
pycparser==2.20
pytz==2021.1
requests==2.26.0
shortuuid==1.0.1
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.6
validators==0.18.2
whitenoise==5.3.0
